Core Functionality Complete - Implemented functionality to load and save a roster - Refactored some of the javascript code to make it more readable
At this point everything is complete to the point that it is operational. Now I just have to go into the api and make sure that is cleaned up and proper input validation is put in place. I also want to put in some input validation on the client side to save some unnecessary trips to the server.