Bug fix sorta
I was encountering a bug on my macbook which caused the excel file to download as an empty txt. It actually took me quite a while to fix this but after further investigation, I for some reason
had the http cors header set to 'no-cors' for whatever reason. My desktop did not have that set so I deleted it and everything works now. After this commit both versions on my mac and desktop will be
the same so there should not be any issues anymore